--- __tests__/components/Send.timers.test.js
@@ -1,232 +1,0 @@
-import React from 'react'
-import { Provider } from 'react-redux'
-import thunk from 'redux-thunk'
-import configureStore from 'redux-mock-store'
-import { shallow, mount } from 'enzyme'
-import { SHOW_NOTIFICATION, HIDE_NOTIFICATION } from '../../app/modules/notification'
-import Send from '../../app/containers/Send'
-import axios from 'axios'
-import MockAdapter from 'axios-mock-adapter'
-
-const axiosMock = new MockAdapter(axios)
-axiosMock.onAny().reply(200)
-
-const initialState = {
-  account: {
-    wif: 'L4SLRcPgqNMAMwM3nFSxnh36f1v5omjPg3Ewy1tg2PnEon8AcHou',
-    address: 'AWy7RNBVr9vDadRMK9p7i7Z1tL7GrLAxoh'
-  },
-  metadata: {
-    network: 'TestNet'
-  },
-  wallet: {
-    Neo: 5,
-    Gas: 1
-  },
-  transactions: {
-    selectedAsset: 'Neo'
-  },
-  dashboard: {
-    confirmPane: true
-  }
-}
-
-const setup = (state = initialState, shallowRender = true) => {
-  const store = configureStore([thunk])(state)
-
-  let wrapper
-  if (shallowRender) {
-    wrapper = shallow(<Send store={store} />)
-  } else {
-    wrapper = mount(
-      <Provider store={store}>
-        <Send />
-      </Provider>
-    )
-  }
-
-  return {
-    store,
-    wrapper
-  }
-}
-
-const changeAmount = (amount, wrapper) => {
-  const amountField = wrapper.find('input[placeholder="Amount"]')
-  amountField.instance().value = amount
-  amountField.simulate('change')
-}
-
-const changeAddress = (address, wrapper) => {
-  const addressField = wrapper.find('input[placeholder="Where to send the asset (address)"]')
-  addressField.instance().value = address
-  addressField.simulate('change')
-}
-
-jest.useFakeTimers()
-
-describe('sendAsset button fake timer tests', () => {
-  test('sendAsset button is getting called correctly with no address or amount sent', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'Please specify an address and amount',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly with invalid address', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-
-    changeAddress('This-is-an-invalid-address', wrapper)
-    changeAmount(initialState.wallet.Neo, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'The address you entered was not valid.',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly with fractional NEO', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-
-    changeAddress(initialState.account.address, wrapper)
-    changeAmount(initialState.wallet.Neo + 0.5, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'You cannot send fractional amounts of Neo.',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly with no NEO amount', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-
-    changeAddress(initialState.account.address, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'Please specify an address and amount',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly with not enough NEO', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-
-    changeAddress(initialState.account.address, wrapper)
-    changeAmount(initialState.wallet.Neo + 1, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'You do not have enough NEO to send.',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly with negative NEO', (done) => {
-    const { wrapper, store } = setup(initialState, false)
-
-    changeAddress(initialState.account.address, wrapper)
-    changeAmount(-1, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'You cannot send negative amounts of an asset.',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-
-  test('sendAsset button is getting called correctly for without enough Gas', (done) => {
-    const gasState = Object.assign({}, initialState, { transactions: { selectedAsset: 'Gas' } })
-    const { wrapper, store } = setup(gasState, false)
-
-    changeAddress(initialState.account.address, wrapper)
-    changeAmount(initialState.wallet.Gas + 1, wrapper)
-
-    wrapper.find('#doSend').simulate('click')
-
-    jest.runAllTimers()
-    const actions = store.getActions()
-    expect(actions.length === 2).toEqual(true)
-    expect(actions[0]).toEqual({
-      type: SHOW_NOTIFICATION,
-      payload: {
-        message: 'You do not have enough GAS to send.',
-        type: 'ERROR'
-      }
-    })
-    expect(actions[1]).toEqual({
-      type: HIDE_NOTIFICATION
-    })
-    done()
-  })
-})
